fix: invalid concurrency number parsing
When someone specified the concurrency as a number with decimals, it was parsed to an integer without any errors. After this fix, it will throw an error (in the before mentioned case) telling you to specify a positive integer instead.

Breaking change: when you use a decimal number to specify the concurrency amount, it will result in an error being thrown. Use a positive integer instead.