ci: change dependency audit to prod only
When pnpm audit fails on a dev dependency, it is usually not a problem. But it currently does block Dependabot PR's which should update dependencies and fixes most of the time the failed audit. Therefore, I chose to check only prod dependencies. The dev dependencies are already in some other checks, which doesn't block PR's and can still be monitored in that way.